# backend/services/question_service/docs/api/OPENAPI_EG_DOCS.yaml
openapi: 3.0.3
info:
  title: Question Service API (PeerPrep)
  version: 0.3.0
  description: |
    Microservice to manage and serve coding interview questions.

    Key behaviors:
    - Questions have:
      - problem statement in Markdown (`body_md`)
      - sanitized HTML (`body_html`) with signed image URLs on read
      - attachments stored in S3, surfaced through presigned/signed URLs
      - difficulty, topics, version, status
      - execution resources:
        - starter code (currently python only)
        - test cases (sample vs hidden)

    - Topics are managed separately via `/admin/topics` and referenced by slug.

    - `/select` is idempotent for a `matching_id` for ~10 minutes.

    - Admin endpoints require `admin` role (via `x-role: admin` in dev or RBAC'ed JWT in prod).

    - Service-to-service calls (like the matching worker) may use `x-role: service`.
servers:
  - url: http://localhost:3000
    description: Local dev
  - url: https://api.example.com/question-service
    description: Production (placeholder)
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Production calls should include a Bearer token.
        In local/dev we also allow `x-role: admin` or `x-role: service`.
  parameters:
    Difficulty:
      name: difficulty
      in: query
      description: Optional difficulty filter
      schema:
        $ref: '#/components/schemas/Difficulty'
    Page:
      name: page
      in: query
      description: 1-based page number
      schema:
        type: integer
        minimum: 1
        default: 1
    Size:
      name: size
      in: query
      description: Page size (max 50)
      schema:
        type: integer
        minimum: 1
        maximum: 50
        default: 20
    Q:
      name: q
      in: query
      description: Full-text search (title/body)
      schema:
        type: string
    Topics:
      name: topics
      in: query
      description: 'Comma-separated topic slugs (e.g., `arrays,graphs`).'
      schema:
        type: string
    Highlight:
      name: highlight
      in: query
      description: Include short snippet highlighting matches (only applies if `q` is provided)
      schema:
        type: boolean
        default: false
  schemas:
    Difficulty:
      type: string
      enum: [Easy, Medium, Hard]
    Status:
      type: string
      enum: [draft, published, archived]
    Error:
      type: object
      properties:
        error:
          type: string
          example: not_found
        message:
          type: string
          example: Question not found
      required:
        - error
    Attachment:
      type: object
      description: Attachment metadata stored on the question row.
      properties:
        object_key:
          type: string
          example: questions/two-sum/20251031/abc123-diagram.png
        mime:
          type: string
          example: image/png
        alt:
          type: string
          nullable: true
          example: diagram
        byte_size:
          type: integer
          minimum: 0
          nullable: true
        width:
          type: integer
          minimum: 1
          nullable: true
        height:
          type: integer
          minimum: 1
          nullable: true
      required:
        - object_key
        - mime
    TopicTag:
      type: object
      description: Topic metadata (used in read models).
      additionalProperties: false
      properties:
        slug:
          type: string
          example: arrays
          pattern: '^[a-z0-9]+(?:-[a-z0-9]+)*$'
        color_hex:
          type: string
          example: '#3b82f6'
          pattern: '^#(?:[0-9a-fA-F]{6}|[0-9a-fA-F]{3})$'
      required:
        - slug
        - color_hex
...
    TestCaseWrite:
      type: object
      description: |
        Test case provided by admin on create/patch.
      properties:
        visibility:
          type: string
          enum: [sample, hidden]
          description: |
            - sample -> visible to users
            - hidden -> internal only, not exposed publicly
        input_data:
          type: string
          description: Raw input (JSON, stdin args, etc.)
        expected_output:
          type: string
          description: Expected output (JSON, stdout, etc.)
        ordinal:
          type: integer
          description: Stable order for runner / UI
          example: 0
      required:
        - visibility
        - input_data
        - expected_output
    TestCaseAdminRead:
      type: object
      description: Test case returned to admin/service.
      properties:
        name:
          type: string
          example: case-0
        visibility:
          type: string
          enum: [sample, hidden]
        input:
          type: string
          example: '[[1,2],[3,4]]'
        expected:
          type: string
          example: '[[3,1],[4,2]]'
        ordinal:
          type: integer
          example: 0
      required:
        - visibility
        - input
        - expected
        - ordinal
...
    QuestionBase:
      type: object
      properties:
        id:
          type: string
          example: rotate-image-in-place
        title:
          type: string
        body_md:
          type: string
          description: |
            Markdown body.

            Inline images in authoring can refer to internal objects using:
            `pp://<object_key>`.

            These keys are finalized on create/patch:
            `staging/...` → `questions/{slug}/...`.
        difficulty:
          $ref: '#/components/schemas/Difficulty'
        topics:
          type: array
          description: |
            Topics for this question.

            On READ (admin/public):
            - array of `{ slug, color_hex }`.

            On WRITE (create/patch):
            - just an array of slugs.
          items:
            $ref: '#/components/schemas/TopicTag'
        attachments:
          type: array
          items:
            $ref: '#/components/schemas/Attachment'
        status:
          $ref: '#/components/schemas/Status'
        version:
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - id
        - title
        - body_md
        - difficulty
        - status
        - version
    QuestionCreate:
      type: object
      description: |
        Create a new draft question.

        Flow:
        1. Admin uploads assets to S3 via /admin/attachments/sign-upload.
        2. Admin references those `staging/...` keys in `attachments` and
           `body_md` using `pp://<object_key>`.
        3. Service persists draft, copies any staging objects into
           `questions/{slug}/...`, rewrites `body_md`, and stores final attachments.
        4. Service also upserts execution resources (starter code + test cases).
      properties:
        title:
          type: string
          example: Rotate Image In-Place
        body_md:
          type: string
          example: |
            Given n x n matrix, rotate 90°.
            ![diagram](pp://staging/dev-user/.../diagram.png)
        difficulty:
          $ref: '#/components/schemas/Difficulty'
        topics:
          type: array
          description: |
            Array of topic slugs (must already exist, created via /admin/topics).
          items:
            type: string
          example:
            - arrays
            - matrix
            - in-place
        attachments:
          type: array
          description: |
            Attachments for the question.

            Each object_key may still be `staging/...`; the service will finalize it.
          items:
            type: object
            required:
              - object_key
              - mime
            properties:
              object_key:
                type: string
                example: staging/dev-user/01ABC.../20251101/01XYZ-diagram.png
              mime:
                type: string
                example: image/png
              alt:
                type: string
                example: diagram
        starter_code_python:
          type: string
          description: |
            Starter code for the candidate (Python).
            Stored in `question_python_starter`.
          example: |
            def rotate_image(matrix: list[list[int]]) -> None:
                # TODO: rotate in place
                pass
        test_cases:
          type: array
          description: |
            Complete list of test cases to associate with this draft.
            Both `sample` and `hidden` are allowed.
            This replaces any existing test cases.
          items:
            $ref: '#/components/schemas/TestCaseWrite'
      required:
        - title
        - body_md
        - difficulty
...
    QuestionAdminRead:
      allOf:
        - $ref: '#/components/schemas/QuestionBase'
        - type: object
          description: |
            Admin/internal full view (works for any status):
            - question metadata
            - sanitized body_html with signed URLs
            - ALL test cases (including hidden)
            - starter_code_python
          properties:
            body_html:
              type: string
              description: Sanitized HTML derived from body_md, with signed image URLs.
            starter_code_python:
              type: string
              nullable: true
              description: Pulled from `question_python_starter`.
            test_cases:
              type: array
              description: All test cases, including `hidden`.
              items:
                $ref: '#/components/schemas/TestCaseAdminRead'
          required:
            - body_html
...
    QuestionListItem:
      allOf:
        - $ref: '#/components/schemas/QuestionBase'
        - type: object
          description: Row in a paginated list.
          properties:
            body_md:
              type: string
              nullable: true
              description: body_md may be truncated or omitted in list responses.
            snippet:
              type: string
              nullable: true
              description: Highlighted excerpt if `q` and `highlight=true`.
    QuestionListResponse:
      type: object
      properties:
        page:
          type: integer
        page_size:
          type: integer
        total:
          type: integer
        items:
          type: array
          items:
            $ref: '#/components/schemas/QuestionListItem'
      required:
        - page
        - page_size
        - total
        - items
...
security:
  - bearerAuth: []
...
paths:
...
  /admin/questions:
    get:
      tags:
        - Admin
      summary: List questions (any status)
      description: |
        Admin view over all questions (`draft`, `published`, `archived`) with filters.
        Result items are summaries (no test cases inline).
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
        - $ref: '#/components/parameters/Difficulty'
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/Status'
        - $ref: '#/components/parameters/Topics'
        - $ref: '#/components/parameters/Q'
        - $ref: '#/components/parameters/Highlight'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuestionListResponse'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
      security:
        - bearerAuth: []
      x-roles:
        - admin

    post:
      tags:
        - Admin
      summary: Create a new draft question
      description: |
        - Creates the initial `questions` row with `status = draft`.
        - Generates a unique slug (`id`).
        - Finalizes any `staging/...` attachments (copies them under `questions/{id}/...`)
          and rewrites `body_md` accordingly.
        - Persists starter code (python) into `question_python_starter`.
        - Replaces any existing test cases for that question with the provided array.
        - Rewrites topics → `question_topics` join table (FK to `topics.slug`).
        - Returns full admin view (including hidden tests and starter code).

        IMPORTANT:
        - All slugs in `topics` MUST already exist in `/admin/topics`.
        - If you reference an unknown topic slug, the service SHOULD 400
          with `unknown_topics` (FK failures will otherwise surface as 500).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QuestionCreate'
      responses:
        '201':
          description: Created draft
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuestionAdminRead'
        '400':
          description: Validation error (missing title, bad topic slug, etc.)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
      security:
        - bearerAuth: []
      x-roles:
        - admin
...
  /questions:
    get:
      tags:
        - Read
      summary: List published questions
      description: |
        Returns a paginated list of **published** questions only.

        Filters:
        - `difficulty`
        - `topics` (comma-separated slugs)
        - `q` (basic search / FTS)
        - `highlight` (include `snippet`)
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
        - $ref: '#/components/parameters/Difficulty'
        - $ref: '#/components/parameters/Topics'
        - $ref: '#/components/parameters/Q'
        - $ref: '#/components/parameters/Highlight'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuestionListResponse'
        '401':
          description: Unauthorized (if anonymous read is disabled)
        '403':
          description: Forbidden
...